Code Cleanup and TODO markings
