+++ intel_nn_hal/ngraph_creator/src/NgraphNodes.cpp
@@ -4,9 +4,7 @@ namespace android {
 namespace hardware {
 namespace neuralnetworks {
 namespace nnhal {
-NgraphNodes::NgraphNodes(size_t size) : INVALID_STRING("Invalid Node") {
-    mOperationOutputs.reserve(size);
-}
+NgraphNodes::NgraphNodes(size_t size) { mOperationOutputs.reserve(size); }
 
 void NgraphNodes::addInputParam(size_t index, std::shared_ptr<ngraph::opset3::Parameter> inParam) {
     mInputParamsMap[index] = inParam;
@@ -19,6 +17,7 @@ ngraph::Output<ngraph::Node> NgraphNodes::getOperationOutput(size_t index) {
 }
 
 void NgraphNodes::setResultNode(size_t outputIndex) {
+    // TODO: Replace Convert with Result
     // Using Convert instead of Result due to crash with libinference_engine_legacy
     std::shared_ptr<ngraph::Node> resultNode = std::make_shared<ngraph::opset3::Convert>(
         mOperationOutputs[outputIndex], ngraph::element::f32);
@@ -26,6 +25,8 @@ void NgraphNodes::setResultNode(size_t outputIndex) {
 }
 
 const std::string& NgraphNodes::getNodeName(size_t index) {
+    // The getNodeName is expected to be called only for Inputs and Outputs.
+    // Hence, scan through mInputParamsMap and mResultsMap to identify a valid node name.
     if (mResultsMap.find(index) != mResultsMap.end()) return mResultsMap[index]->get_name();
     if (mInputParamsMap.find(index) != mInputParamsMap.end())
         return mInputParamsMap[index]->get_name();
@@ -34,10 +35,11 @@ const std::string& NgraphNodes::getNodeName(size_t index) {
 
 std::shared_ptr<ngraph::Function> NgraphNodes::generateGraph() {
     std::vector<std::shared_ptr<ngraph::Node>> resultNodes;
-    resultNodes.reserve(mResultsMap.size() + 1);
+    resultNodes.reserve(mResultsMap.size());
     for (auto const& temp : mResultsMap) resultNodes.push_back(temp.second);
-    resultNodes.push_back(std::make_shared<ngraph::opset3::Concat>(
-        resultNodes, 1));  // Dummy Concat to join the network
+    // TODO: Remove the Dummy Concat
+    // Dummy Concat to join the disconnected graph(ssd_mobilenet Obj Det with only Concat)
+    resultNodes.push_back(std::make_shared<ngraph::opset3::Concat>(resultNodes, 1));
     ngraph::ParameterVector inputParams;
     inputParams.reserve(mInputParamsMap.size());
     for (auto const& temp : mInputParamsMap) inputParams.push_back(temp.second);
